requirements.txt
pyserini
faiss-cpu
transformers
tqdm
jieba-0.42.1
gensim